ewan

+++ app/Models/ApplicantDetail.php
@@ -9,8 +9,8 @@ class ApplicantDetail extends Model
 {
     use HasFactory;
     protected $fillable = ['login_id','information_id','educ','Applyingfor','resume'];
-    public $keyType = 'bigint';
-public $primaryKey = 'applicant_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'applicant_id';
 
     public function UserDetail(){
         return $this->hasOne(UserDetail::class,'information_id', 'information_id');
