+++ app/Models/Attendance.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Model;
 class Attendance extends Model
 {
     use HasFactory;
-    public $keyType = 'bigint';
-public $primaryKey = 'attendance_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'attendance_id';
 
     protected $fillable = ['employee_id','time_in','time_out','attendance_day','attendance_date'];
 }
