+++ app/Models/Bonus.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Model;
 class Bonus extends Model
 {
     use HasFactory;
-    public $keyType = 'bigint';
-public $primaryKey = 'bonus_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'bonus_id';
 
     protected $fillable = ['payrollManager_id','employee_id','bonus_date','bonus_amount'];
 }
