+++ app/Models/CashAdvance.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Model;
 class CashAdvance extends Model
 {
     use HasFactory;
-    public $keyType = 'bigint';
-public $primaryKey = 'cashAdvance_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'cashAdvance_id';
 
     protected $fillable = ['payrollManager_id','employee_id','cash_advance_date','cashAdvance_amount'];
 }
