+++ app/Models/Contributions.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Model;
 class Contributions extends Model
 {
     use HasFactory;
-    public $keyType = 'bigint';
-public $primaryKey = 'contribution_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'contribution_id';
 
     protected $fillable = ['employee_contribution','employer_contribution','add_high','add_low','low_limit','high_limit'];
 }
