+++ app/Models/Message.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Model;
 class Message extends Model
 {
     use HasFactory;
-    public $keyType = 'bigint';
-public $primaryKey = 'message_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'message_id';
 
     protected $fillable = ['sender_id','receiver_id','message'];
 }
