+++ app/Models/Overtime.php
@@ -9,8 +9,8 @@ class Overtime extends Model
 {
     use HasFactory;
 
-    public $keyType = 'bigint';
-public $primaryKey = 'overtime_id';
+    public $keyType = 'integer';
+public $primaryKey  = 'overtime_id';
 
     protected $fillable = ['employee_id','attendance_id','payrollManager_id'];
